@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\PokemonPRNG.MT\PokemonPRNG.MT.csproj" />
+    <ProjectReference Include="..\PokemonPRNG.SFMT.SIMD\PokemonPRNG.SFMT.SIMD.csproj" />
     <ProjectReference Include="..\PokemonPRNG.SFMT\PokemonPRNG.SFMT.csproj" />
   </ItemGroup>
 
